+++ src/test/java/qna/repository/question/QuestionRepositoryTest.java
@@ -1,10 +1,11 @@
-package qna.repository;
+package qna.repository.question;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import qna.domain.Question;
+import qna.repository.QuestionRepository;
 
 import java.util.List;
 import java.util.Optional;
