+++ src/test/java/qna/repository/user/UserRepositoryTest.java
@@ -1,10 +1,11 @@
-package qna.repository;
+package qna.repository.user;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import qna.domain.User;
+import qna.repository.UserRepository;
 
 import java.util.Optional;
